Removing useless code and moving file
